requirements.txt
cudatoolkit==11.3
ipdb
matplotlib
numpy
pytorch
torchvision
